requirements.txt
python-dotenv==1.0.0
github3.py==3.1.0
requests==2.31.0
rich==13.7.0
PyGithub==2.1.1
docker==6.1.3
python-gitlab==3.15.0
scorecard==0.3  # Using the latest available version
bandit>=1.7.5
safety>=2.3.5
